updating broken tests to use new services
