+++ intake/exceptions.py
@@ -1,0 +1,10 @@
+class JinjaNotInitializedError(Exception):
+    pass
+
+
+class DuplicateTemplateError(Exception):
+    pass
+
+
+class FrontAPIError(Exception):
+    pass
