+++ intake/submission_bundler.py
@@ -4,6 +4,7 @@ from intake import (
     models as intake_models
 )
 from user_accounts.models import Organization
+import intake.services.bundles as BundlesService
 
 
 def is_the_weekend():
@@ -32,7 +33,7 @@ class OrganizationBundle:
         """If there are new submissions, make a new bundle
         If the submissions are the same as the last bundle, use the last bundle
         """
-        return intake_models.ApplicationBundle.create_with_submissions(
+        return BundlesService.create_bundle_from_submissions(
             submissions=self.submissions,
             organization=self.organization,
         )
